migrated mast cell creation, made various tweaks
