+++ client/components/GameFunctions/createFunctions/index.js
@@ -4,7 +4,11 @@ import { scoreAndStars } from './lifeStars';
 import { NPCCells } from './cells'
 import { epithelialCells, makeEpithelialCell, epithelialCellCollision } from './epithelialCells'
 import { tCells, makeTCell } from './tCells'
+import { mastCells, makeMastCell, activate } from './mastCells'
 
-export { players, keyboardControls, scoreAndStars, NPCCells, 
+export { 
+  players, keyboardControls, scoreAndStars, NPCCells, 
   epithelialCells, makeEpithelialCell, epithelialCellCollision,
-  tCells, makeTCell }
+  tCells, makeTCell,
+  mastCells, makeMastCell, activate
+}
